+++ wke/wkeChromeClient.cpp
@@ -573,7 +573,7 @@ WebCore::Page* ChromeClient::createWindow(WebCore::Frame*, const WebCore::FrameL
 {
     // cexer 实现新窗口控制
     wke::CWebViewHandler& handler = m_webView->m_handler;
-    if (!handler.newWindowCallback)
+    if (!handler.createViewCallback)
         return m_webView->page();
 
     wkeNavigationType type = (wkeNavigationType)action.type();
@@ -590,7 +590,7 @@ WebCore::Page* ChromeClient::createWindow(WebCore::Frame*, const WebCore::FrameL
     windowFeatures.toolBarVisible = features.toolBarVisible;
     windowFeatures.fullscreen = features.fullscreen;
 
-    wke::CWebView* createdWebView = handler.newWindowCallback(m_webView, handler.newWindowCallbackParam, type, &url, &windowFeatures);
+    wke::CWebView* createdWebView = handler.createViewCallback(m_webView, handler.createViewCallbackParam, type, &url, &windowFeatures);
     if (!createdWebView)
         return m_webView->page();