+++ wkexe/main.c
@@ -1,0 +1,15 @@
+
+#include "app.h"
+
+
+int APIENTRY WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPWSTR lpCmdLine, int nCmdShow)
+{
+    wkeInitialize();
+    {
+        Application app;
+        RunApplication(&app);
+    }
+    wkeFinalize();
+
+    return 0;
+}